refactor(agent): move standalone package
