+++ agent/agent.go
@@ -5,16 +5,17 @@ import (
 
 	"github.com/docker/docker/client"
 	"github.com/pkg/errors"
-	"github.com/portainer/portainer-updater/update"
+	"github.com/portainer/portainer-updater/dockerstandalone"
 	"github.com/rs/zerolog/log"
 )
 
-type AgentUpdateCommand struct {
-	Image      string `arg:"" help:"Image of the agent to upgrade to. e.g. portainer/agent:latest" name:"image"`
+type AgentCommand struct {
+	EnvType    string `kong:"help='The environment type',default='standalone',enum='standalone,nomad'"`
+	Image      string `arg:"" help:"Image of the agent to upgrade to. e.g. portainer/agent:latest" name:"image" default:"portainer/agent:latest"`
 	ScheduleId string `arg:"" help:"Schedule ID of the agent to upgrade to. e.g. 1" name:"schedule-id"`
 }
 
-func (r *AgentUpdateCommand) Run() error {
+func (r *AgentCommand) Run() error {
 	dockerCli, err := client.NewClientWithOpts(client.FromEnv, client.WithAPIVersionNegotiation())
 	if err != nil {
 		log.Fatal().Err(err).Msg("Unable to initialize Docker client")
@@ -26,16 +27,16 @@ func (r *AgentUpdateCommand) Run() error {
 		Str("image", r.Image).
 		Str("schedule-id", r.ScheduleId).
 		Msg("Updating Portainer agent")
-	oldContainer, err := findContainer(ctx, dockerCli)
+	oldContainer, err := dockerstandalone.FindAgentContainer(ctx, dockerCli)
 	if err != nil {
 		return errors.WithMessage(err, "failed finding container id")
 	}
 
-	if oldContainer.Labels != nil && oldContainer.Labels[update.UpdateScheduleIDLabel] == r.ScheduleId {
+	if oldContainer.Labels != nil && oldContainer.Labels[dockerstandalone.UpdateScheduleIDLabel] == r.ScheduleId {
 		log.Info().Msg("Agent already updated")
 
 		return nil
 	}
 
-	return update.Update(ctx, dockerCli, oldContainer.ID, r.Image, r.ScheduleId)
+	return dockerstandalone.Update(ctx, dockerCli, oldContainer.ID, r.Image, r.ScheduleId)
 }
