+++ cli/types.go
@@ -6,7 +6,7 @@ import (
 )
 
 var CLI struct {
-	LogLevel    log.Level                `kong:"help='Set the logging level',default='INFO',enum='DEBUG,INFO,WARN,ERROR',env='LOG_LEVEL'"`
-	PrettyLog   bool                     `kong:"help='Whether to enable or disable colored logs output',default='false',env='PRETTY_LOG'"`
-	AgentUpdate agent.AgentUpdateCommand `cmd:"" help:"Update an existing Portainer agent container."`
+	LogLevel  log.Level          `kong:"help='Set the logging level',default='INFO',enum='DEBUG,INFO,WARN,ERROR',env='LOG_LEVEL'"`
+	PrettyLog bool               `kong:"help='Whether to enable or disable colored logs output',default='false',env='PRETTY_LOG'"`
+	Agent     agent.AgentCommand `cmd:"" help:"Update an existing Portainer agent container."`
 }
